Enable tests in Travis, add some exclusions
Enabling tests in the Travis CI build to give us higher confidence in
proposed changes. The Travis environment seems to introduce flakiness or
even total failure for some tests, so this adds test retries for some
tests and total exclusion for other tests in order to alleviate this.
The 'ci-flaky' test group is excluded for all test tasks in the CI
environment, though these tests should be addressed soon.